docs: update docgen
