--- packages/docs/src/docgen-watch.ts
@@ -1,3 +1,0 @@
-import { run } from "./run";
-
-run({ watch: true });
